properly split code between headers and source files. Initial commit for window.h/cpp and drawable.h/cpp; they're extracted from application.h/cpp
