+++ Makefile
@@ -17,6 +17,8 @@ tga_image.cpp                          \
 textbox.cpp                            \
 level2/input.cpp                       \
 level2/application.cpp                 \
+level2/drawable.cpp                    \
+level2/window.cpp                      \
 level2/drawing_context_proxy.cpp       \
 level2/widgets/label.cpp               \
 level2/widgets/button.cpp              \
