+++ level2/draw_area.h
@@ -28,6 +28,7 @@
 #pragma once
 
 #include <utility>
+#include <vector>
 #include <list>
 #include "point.h"
